backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyMuPDF
markdown==3.5.1
html2text==2020.1.16
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
faiss-cpu==1.12.0
numpy
scikit-learn
python-dotenv==1.0.0
aiofiles==23.2.0
google-genai